requirements.txt
dal
attrs==18.1.0
chardet==3.0.4
coverage==4.5.1
Django==2.0.1
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-csvimport==2.11
django-debug-toolbar==1.9.1
django-dynamic-formsets==0.0.8
django-extensions==2.0.7
-e git+https://github.com/coleifer/django-generic-m2m.git@83b90cd38275b04413f944fe37e4446e4fe09bf6#egg=django_generic_m2m
django-gm2m==0.6.1
django-js-asset==1.1.0
-e git+https://github.com/theatlantic/django-nested-admin.git@4a511db92549d7c17c3499638060f0761dcb414e#egg=django_nested_admin
django-querysetsequence==0.8
django-searchable-select==1.5.0
django-tables2==1.19.0
django-tagging==0.4.6
django-taggit==0.22.2
et-xmlfile==1.0.1
jdcal==1.4
mock==2.0.0
more-itertools==4.1.0
numpy==1.14.2
openpyxl==2.5.4
pbr==4.0.2
Pillow==5.0.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-splinter==1.8.5
python-monkey-business==1.0.0
pytz==2017.3
selenium==3.12.0
six==1.11.0
splinter==0.8.0
sqlparse==0.2.4
tenacity==4.12.0
wget==3.2
whitenoise==3.3.1
